moved transformations and add doc for rayleigh properties
